+++ packages/ethers5/src/utils/defaultConfig.ts
@@ -5,10 +5,6 @@ import { CoinbaseWalletSDK } from '@coinbase/wallet-sdk'
 export interface ConfigOptions {
   enableEIP6963?: boolean
   enableCoinbase?: boolean
-  /**
-   * Use enableEIP6963 to show all injected wallets
-   * @deprecated
-   */
   enableInjected?: boolean
   rpcUrl?: string
   defaultChainId?: number
@@ -16,11 +12,38 @@ export interface ConfigOptions {
 }
 
 export function defaultConfig(options: ConfigOptions) {
-  const { enableEIP6963 = true, enableCoinbase = true, metadata, rpcUrl, defaultChainId } = options
+  const {
+    enableEIP6963 = true,
+    enableInjected = true,
+    enableCoinbase = true,
+    metadata,
+    rpcUrl,
+    defaultChainId
+  } = options
 
+  let injectedProvider: Provider | undefined = undefined
   let coinbaseProvider: Provider | undefined = undefined
 
   const providers: ProviderType = { metadata }
+
+  function getInjectedProvider() {
+    if (injectedProvider) {
+      return injectedProvider
+    }
+
+    if (typeof window === 'undefined') {
+      return undefined
+    }
+
+    if (!window.ethereum) {
+      return undefined
+    }
+
+    //  @ts-expect-error window.ethereum satisfies Provider
+    injectedProvider = window.ethereum
+
+    return injectedProvider
+  }
 
   function getCoinbaseProvider() {
     if (coinbaseProvider) {
@@ -45,6 +68,10 @@ export function defaultConfig(options: ConfigOptions) {
 
   if (enableCoinbase && rpcUrl && defaultChainId) {
     providers.coinbase = getCoinbaseProvider()
+  }
+
+  if (enableInjected) {
+    providers.injected = getInjectedProvider()
   }
 
   if (enableEIP6963) {